Create PS Sas token
This script generates a SAS token for the container "mycontainer" in the storage account "mystorageaccount". The SAS token allows read, write, delete, and list permissions on the container and expires in 1 hour.

You can modify this script to set different properties for the SAS token, such as the expiry time or the allowed permissions.