Create PS AAFD Auth for Storage account
This script assigns the "Storage Blob Data Contributor" role to a user for the container "mycontainer" in the storage account "mystorageaccount". The user will be able to access the container using Azure AD authentication.

You can modify this script to assign different roles or to assign roles to different users or groups.