Create PS Import Job
This script creates an import job to import a VHD file from a SAS URI into the container "mycontainer" in the storage account "mystorageaccount". The import job will overwrite any existing blobs in the container and use 2 threads to transfer the data.

You can modify this script to set different options for the import job, such as the import type or the thread count.