Create PS Az Copy
This script uses AzCopy to copy all files from the local folder "C:\myfolder" to the container "mycontainer" in the storage account "mystorageaccount". It uses the storage account access key for authentication and overwrites any existing blobs in the container.

You can modify this script to set different options for the copy operation, such as the copy type or the number of threads.